Collection#[] - request only item if accessed with entry_filter
To do so, add #request_entry method to ItemCollection and UserCollection.
Also, add a new 'initialized' param to #entry and update API#item, #user
to support it. Add new webmock stubs to the tests to support requests for
individual items.

Also, add missing docs on child class methods to Collection.